+++ src/main/java/org/archive/url/GoogleURLCanonicalizer.java
@@ -1,7 +1,0 @@
-package org.archive.url;
-
-/**
- * @deprecated use {@link BasicURLCanonicalizer}
- */
-public class GoogleURLCanonicalizer extends BasicURLCanonicalizer {
-}