+++ src/test/java/org/archive/url/IAURLCanonicalizerTest.java
@@ -10,7 +10,7 @@ public class IAURLCanonicalizerTest {
 
 	@Test
 	public void testFull() throws URISyntaxException {
-		IAURLCanonicalizer iaC = new IAURLCanonicalizer(new DefaultIACanonicalizerRules());
+		IAURLCanonicalizer iaC = new IAURLCanonicalizer(new AggressiveIACanonicalizerRules());
 		compCan(iaC,"http://www.archive.org:80/","http://archive.org/");
 		compCan(iaC,"https://www.archive.org:80/","https://archive.org:80/");
 		compCan(iaC,"http://www.archive.org:443/","http://archive.org:443/");
@@ -63,7 +63,7 @@ public class IAURLCanonicalizerTest {
 
 	@Test
 	public void testStripSessionId() throws URISyntaxException {
-	    IAURLCanonicalizer iaC = new IAURLCanonicalizer(new DefaultIACanonicalizerRules());
+	    IAURLCanonicalizer iaC = new IAURLCanonicalizer(new AggressiveIACanonicalizerRules());
 	    compCan(iaC,
                 "http://www.nsf.gov/statistics/sed/2009/SED_2009.zip?CFID=14387305&CFTOKEN=72942008&jsessionid=f030eacc7e49c4ca0b077922347418418766",
                 "http://nsf.gov/statistics/sed/2009/sed_2009.zip?jsessionid=f030eacc7e49c4ca0b077922347418418766");